dgify tests
