+++ docs/docs/guides/test/running-a-subset-of-asset-checks.md
@@ -23,7 +23,7 @@ Inside the body of the function, we can use `AssetCheckExecutionContext.selected
 
 As we don't know in advance which checks will be executed, we explicitly `yield` each asset check result that we're expected to create:
 
-<CodeExample path="docs_snippets/docs_snippets/concepts/assets/asset_checks/subset_multi_asset_check.py" />
+<CodeExample path="docs_snippets/docs_snippets/concepts/assets/asset_checks/subset_multi_asset_check.py" title="src/my_project/defs/asset_checks.py" />
 
 ## Subsetting checks in @multi_assets
 
@@ -31,7 +31,7 @@ When using [multi-assets](/guides/build/assets/defining-assets#multi-asset), Dag
 
 In the following example, we only want to execute a check when the `multi_asset_piece_1` asset produced by the `multi_asset_1_and_2` multi-asset is materialized:
 
-<CodeExample path="docs_snippets/docs_snippets/concepts/assets/asset_checks/subset_check_multi_asset.py" />
+<CodeExample path="docs_snippets/docs_snippets/concepts/assets/asset_checks/subset_check_multi_asset.py" title="src/my_project/defs/asset_checks.py" />
 
 Let's review what we did to accomplish this:
