+++ examples/docs_snippets/docs_snippets/guides/data-assets/quality-testing/asset-checks/assets.py
@@ -1,0 +1,1 @@
+def orders(): ...
