+++ examples/docs_snippets/docs_snippets/guides/quality-testing/unit-testing-assets-and-ops/asset-context.py
@@ -1,10 +1,6 @@
+from my_project.defs.assets import loaded_file
+
 import dagster as dg
-
-
-@dg.asset(partitions_def=dg.DailyPartitionsDefinition("2024-01-01"))
-def loaded_file(context: dg.AssetExecutionContext) -> str:
-    with open(f"path_{context.partition_key}.txt") as file:
-        return file.read()
 
 
 # highlight-start
