+++ examples/docs_snippets/docs_snippets/guides/quality-testing/unit-testing-assets-and-ops/asset-dependency.py
@@ -1,15 +1,4 @@
-import dagster as dg
-
-
-@dg.asset
-def loaded_file() -> str:
-    with open("path.txt") as file:
-        return file.read()
-
-
-@dg.asset
-def processed_file(loaded_file: str) -> str:
-    return loaded_file.strip()
+from my_project.defs.assets import processed_file
 
 
 # highlight-start
