+++ examples/docs_snippets/docs_snippets/guides/quality-testing/unit-testing-assets-and-ops/op-context.py
@@ -1,10 +1,6 @@
+from my_project.defs.assets import load_file
+
 import dagster as dg
-
-
-@dg.op
-def load_file(context: dg.OpExecutionContext) -> str:
-    with open(f"path_{context.partition_key}.txt") as file:
-        return file.read()
 
 
 # highlight-start
