+++ examples/docs_snippets/docs_snippets/guides/quality-testing/unit-testing-assets-and-ops/op-no-argument.py
@@ -1,10 +1,4 @@
-import dagster as dg
-
-
-@dg.op
-def load_file() -> str:
-    with open("path.txt") as file:
-        return file.read()
+from my_project.defs.assets import load_file
 
 
 # highlight-start
